Extract HSL color space into sub-module
